requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
twilio
mftool
python-dotenv
